requirements.txt
bidict==0.22.1
blinker==1.7.0
chardet==5.2.0
click==8.1.7
colorama==0.4.6
DateTime==5.4
distlib==0.3.8
dominate==2.9.1
filelock==3.13.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
itsdangerous==2.1.2
MarkupSafe==2.1.3
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
reportlab==4.0.9
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
virtualenv==20.25.0
visitor==0.1.3
Werkzeug==3.0.1
wsproto==1.2.0
zope.interface==6.1
psycopg2==2.9.9